Refactor planterbox into modules
Cleanly split related code; present a straightforward export interface in __init__.py
